ｈｐが減るように実装 無敵時間もおまけで実装

+++ Hide/Player.h
@@ -66,7 +66,7 @@ protected:
 	double angle;//カーソルの傾き
 	int invincible;//無敵時間
 
-	float preY; //前のフレームのvelocityYを保持
+	int preY; //前のフレームのvelocityYを保持
 	int interval;//星の発射間隔
 	bool knockback_status;//ノックバック中か
 	std::unique_ptr<StarManager> starmanager;
